pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = "Projects common Python interpreter"
authors = ["Alexander Nesterovich"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Flask = "^2.1.2"
Flask-Admin = "^1.6.0"
Flask-Caching = "^2.0.1"
Flask-DebugToolbar = "^0.11.0"
Flask-Mail = "^0.9.1"
Flask-Sijax = "^0.4.1"
Flask-Testing = "^0.8.1"
Jinja2 = "^3.0.3"
PyMySQL = "^1.0.2"
SQLAlchemy = "^1.4.29"
SQLAlchemy-Utils = "^0.38.2"
WebTest = "^3.0.0"
alembic = "^1.7.5"
email-validator = "^1.1.3"
mongoengine = "^0.23.1"
#mysqlclient = "^2.1.0"
passlib = "^1.7.4"
progressbar2 = "^4.0.0"
pytest = "^7.0.0"
pytest-clarity = "^1.0.1"
pytest-cov = "^3.0.0"
python-dateutil = "^2.8.2"
python-dotenv = "^0.19.2"
redis = "^4.3.1"
requests = "^2.27.1"
sqlacodegen = "^2.3.0"
humanfriendly = "^10.0"
colorama = "^0.4.4"
uvloop = "^0.16.0"
aiohttp = "^3.8.1"
bs4 = "^0.0.1"
#cchardet = "^2.1.7"
codetiming = "^1.3.0"
unsync = "^1.4.0"
Django = "^4.0.6"
matplotlib = "^3.5.2"
numpy = "^1.23.1"
big-O-calculator = "^0.0.9"
pandas = "^1.4.3"
imageio = "^2.20.0"
ipython = "^8.4.0"
keras = "^2.9.0"
tensorflow = "^2.9.1"
jupyter = "^1.0.0"
sympy = "^1.11"
mongita = "^1.1.1"
fastapi = "^0.82.0"
uvicorn = "^0.18.3"
allegro-pl-rest-api = "^2020.3.12"
pipdeptree = "^2.3.3"
yolk = "^0.4.3"
yolk3k = "^0.9"
selenium = "^4.6.0"
py3-validate-email = "^1.0.5"
pytypes = "^1.0b10"
celery = "^5.2.7"
rabbitmq = "^0.2.0"
cookiecutter = "^2.1.1"
coverage = "^6.5.0"
django-coverage-plugin = "^2.0.4"
Sphinx = "^5.3.0"
django-extensions = "^3.2.1"
Werkzeug = "^2.2.2"
django-test-plus = "^2.2.1"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.7.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.6"
wheel = "^0.38.4"
django-model-utils = "^4.3.1"
argon2-cffi = "^21.3.0"
pytz = "^2022.6"
django-celery-results = "^2.4.0"
django-environ = "^0.9.0"
mysql-connector-python = "^8.0.31"
patch-env = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
